editors - move decorateFileEditorLabel to common place
